+++ src/main/java/com/somemore/global/auth/sign/up/SignUpService.java
@@ -4,8 +4,8 @@ import com.somemore.center.usecase.NEWRegisterCenterUseCase;
 import com.somemore.global.auth.oauth.domain.CommonOAuthInfo;
 import com.somemore.global.auth.oauth.registrar.OAuthInfoRegistrar;
 import com.somemore.user.domain.User;
+import com.somemore.user.domain.UserAuthInfo;
 import com.somemore.user.domain.UserRole;
-import com.somemore.user.dto.UserAuthInfo;
 import com.somemore.user.service.RegisterUserService;
 import com.somemore.volunteer.usecase.NEWRegisterVolunteerUseCase;
 import lombok.RequiredArgsConstructor;
