+++ src/test/java/com/somemore/user/service/RegisterUserServiceTest.java
@@ -5,10 +5,10 @@ import com.somemore.global.auth.oauth.domain.OAuthProvider;
 import com.somemore.global.imageupload.service.ImageUploadService;
 import com.somemore.support.IntegrationTestSupport;
 import com.somemore.user.domain.User;
+import com.somemore.user.domain.UserAuthInfo;
 import com.somemore.user.domain.UserCommonAttribute;
 import com.somemore.user.domain.UserRole;
 import com.somemore.user.domain.UserSetting;
-import com.somemore.user.dto.UserAuthInfo;
 import com.somemore.user.repository.user.UserRepository;
 import com.somemore.user.repository.usercommonattribute.UserCommonAttributeRepository;
 import com.somemore.user.repository.usersetting.UserSettingRepository;
